Refactor: Moved similiar code to subpackages, modified all imports
