+++ src/main/java/fundamentals/server/handlers/BuildAllHandler.java
@@ -1,5 +1,6 @@
-package fundamentals.server;
+package fundamentals.server.handlers;
 
+import fundamentals.server.BuildStorage;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
