+++ src/test/java/fundamentals/server/ContinuousIntegrationServerTest.java
@@ -2,7 +2,7 @@ package fundamentals.server;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ContinuousIntegrationServerTest {
